#237 - Fixed: wp_redirect caches with W3 Total Cache
When W3 Total Cache caches the page with a 302 redirect (the default for
wp_redirect) then subsuquent visits to the page will be cached, and
result in the inability to access the Vendor Dashboard.  Most browsers
are smart enough to know that they should ignore that, but some versions
of FireFox (42) do not.  As such, this is the proper fix by changing the
302 redirect to a 303 which is a do-not-cache redirect.